Restructured the compilation process
