+++ exercises/poisson/jacobi_kernel.cpp
@@ -10,7 +10,7 @@ __global__ void compute_jacobi_hip_kernel(double * field_phi_new, double * field
     if ( (i<g->n[0]) and (j < g->n[1]) )
     {
         auto k = (j +1) + (i+1)*(g->n[1]+2);
-
+        
         auto nx = g->n[0];
         auto ny = g->n[1];
         auto dx = g->dx[0];
@@ -28,5 +28,5 @@ void launch_compute_jacobi_hip(double * phi_new_dev, double * phi_old_dev, doubl
     dim3 blockSize(8,8*2,1);
     dim3 gridSize((nx+1)/8,(ny+1)/(8*2),1);
     compute_jacobi_hip_kernel<<<blockSize,gridSize>>>(phi_new_dev, phi_old_dev, rho_dev,g_dev);
-
+    
 }
